backend/requirements.txt
flask>=3.0
flask-cors>=4.0
ultralytics>=8.2
pillow>=10.0
torch>=2.0
